Added a config file
This for now allows the configuration of what input files to use.
Pretty rendering of genders (see german.yml for an example).
And allow ignoring certain keys in the autocomplete.

I will likely later on add the ability to switch between search
algorithms for the autocomplete. Allow completly disabling autocomplete,
etc.